+++ drivers/CoreVideo/include/CoreDSI.hpp
@@ -6,7 +6,7 @@
 #define _LEKA_OS_DRIVER_DSI_H_
 
 #include "LKCoreSTM32HalBase.h"
-#include "interface/CoreDSI.hpp"
+#include "interface/DSI.hpp"
 
 namespace leka {
