+++ drivers/CoreVideo/include/CoreLTDC.hpp
@@ -6,7 +6,7 @@
 #define _LEKA_OS_LIB_LTDC_H_
 
 #include "LKCoreSTM32HalBase.h"
-#include "interface/CoreDSI.hpp"
+#include "interface/DSI.hpp"
 #include "interface/CoreLTDC.hpp"
 
 namespace leka {
