+++ drivers/CoreVideo/include/CoreVideo.hpp
@@ -6,7 +6,7 @@
 #define _LEKA_OS_DRIVER_LKCOREVIDEO_H_
 
 #include "LKCoreSTM32HalBase.h"
-#include "interface/CoreDSI.hpp"
+#include "interface/DSI.hpp"
 #include "interface/CoreFont.hpp"
 #include "interface/CoreGraphics.hpp"
 #include "interface/CoreJPEG.hpp"
