+++ tests/unit/mocks/mocks/leka/CoreDSI.h
@@ -6,7 +6,7 @@
 #define _LEKA_OS_DSI_MOCK_H_
 
 #include "gmock/gmock.h"
-#include "interface/CoreDSI.hpp"
+#include "interface/DSI.hpp"
 
 namespace leka {
